+++ lib/app.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:hive/hive.dart';
+import 'package:statemanager/bloc/bot_bloc.dart';
+import 'package:statemanager/model/bot.dart';
+import 'package:statemanager/ui/home_page.dart';
+
+class App extends StatelessWidget {
+  final Box<Bot> botBox;
+
+  const App({super.key, required this.botBox});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: BlocProvider(
+        create: (_) => BotBloc(botBox),
+        child: const HomePage(),
+      ),
+    );
+  }
+}
